requirements.txt
python-telegram-bot[job-queue]==21.*
psycopg2-binary>=2.9
cachetools>=5.3
